requirements.txt
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dlib==19.24.8
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==79.0.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
